requirements.txt
Flask==0.10.1
Flask-MySQL==1.3
Flask-MySQLdb==0.2.0
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.10.4
argparse==1.2.1
itsdangerous==0.24
mysqlclient==1.3.6
wsgiref==0.1.2
